MSEARCH-117 Move toSafeStream() method to CollectionUtils
- Reduce amount of logs of integration tests
