+++ src/main/java/org/folio/search/service/setter/holding/HoldingsCallNumberComponentsProcessor.java
@@ -1,18 +1,16 @@
 package org.folio.search.service.setter.holding;
 
 import static java.util.stream.Collectors.toSet;
+import static org.folio.search.utils.CollectionUtils.toSafeStream;
 import static org.folio.search.utils.SearchUtils.getEffectiveCallNumber;
-import static org.folio.search.utils.SearchUtils.toSafeStream;
 
 import java.util.Set;
-import lombok.RequiredArgsConstructor;
 import org.apache.commons.lang3.StringUtils;
 import org.folio.search.domain.dto.Instance;
 import org.folio.search.service.setter.FieldProcessor;
 import org.springframework.stereotype.Component;
 
 @Component
-@RequiredArgsConstructor
 public class HoldingsCallNumberComponentsProcessor implements FieldProcessor<Instance, Set<String>> {
 
   @Override
