+++ src/main/java/org/folio/search/service/setter/instance/PublicNotesProcessor.java
@@ -1,6 +1,6 @@
 package org.folio.search.service.setter.instance;
 
-import static org.folio.search.utils.SearchUtils.toSafeStream;
+import static org.folio.search.utils.CollectionUtils.toSafeStream;
 
 import java.util.stream.Stream;
 import org.folio.search.domain.dto.Instance;
