+++ src/main/java/org/folio/search/service/setter/item/EffectiveCallNumberComponentsProcessor.java
@@ -1,8 +1,8 @@
 package org.folio.search.service.setter.item;
 
 import static java.util.stream.Collectors.toSet;
+import static org.folio.search.utils.CollectionUtils.toSafeStream;
 import static org.folio.search.utils.SearchUtils.getEffectiveCallNumber;
-import static org.folio.search.utils.SearchUtils.toSafeStream;
 
 import java.util.Objects;
 import java.util.Set;
