+++ src/main/java/org/folio/search/service/setter/item/ItemPublicNotesProcessor.java
@@ -1,6 +1,6 @@
 package org.folio.search.service.setter.item;
 
-import static org.folio.search.utils.SearchUtils.toSafeStream;
+import static org.folio.search.utils.CollectionUtils.toSafeStream;
 
 import java.util.Collection;
 import java.util.stream.Stream;
